added /urban, added "nqn" setting to /setting, bugfixes

+++ cogs/nonsense.py
@@ -3,6 +3,7 @@ import disnake as discord
 from disnake.ext import commands
 from enum import Enum
 import re
+import os
 import utils
 import random
 import asyncio
@@ -49,23 +50,47 @@ class Nonsense(commands.Cog):
   async def on_message(self, msg):
     if msg.author.bot:
       return
-    reg = ':[a-zA-Z]+:'
-    other = re.split(reg, msg.content)
-    emjs = re.findall(reg, msg.content)
-    content=other[0]
-    for i in range(len(emjs)):
-      myemjs = tuple(filter(lambda emj: emj.name==emjs[i][1:-1], self.bot.emojis))
-      emj = f'<:{myemjs[0].name}:{myemjs[0].id}>' if (any(myemjs) and not other[i].endswith('<')) else emjs[i]
-      content+=emj+other[i+1]
-      
-    if content==msg.content: return
-    if msg.reference and len(msg.content.split())==1:
-      await msg.delete()
-      await self.react.__call__(msg, myemjs[0], msg.reference.resolved)
-    else:
-      webhook = (await utils.Webhook((await self.bot.get_context(msg))))
-      await msg.delete()
-      await webhook.send(content=content, username=msg.author.display_name, avatar_url=msg.author.avatar)
+    if str(msg.guild.id) in db["serversetting"]["nqn"]:
+      reg = ':[a-zA-Z]+:'
+      other = re.split(reg, msg.content)
+      emjs = re.findall(reg, msg.content)
+      content=other[0]
+      for i in range(len(emjs)):
+        myemjs = tuple(filter(lambda emj: emj.name==emjs[i][1:-1], self.bot.emojis))
+        emj = f'<:{myemjs[0].name}:{myemjs[0].id}>' if (any(myemjs) and not other[i].endswith('<')) else emjs[i]
+        content+=emj+other[i+1]
+        
+      if content==msg.content: return
+      if msg.reference and len(msg.content.split())==1:
+        await msg.delete()
+        await self.react.__call__(msg, myemjs[0], msg.reference.resolved)
+      else:
+        webhook = (await utils.Webhook((await self.bot.get_context(msg))))
+        await msg.delete()
+        await webhook.send(content=content, username=msg.author.display_name, avatar_url=msg.author.avatar)
+
+  @commands.slash_command(name = "urban")
+  async def slashurban(inter, query):
+    '''
+    See meaning of term you need
+    Parameters
+    ----------
+    query: Your term here!
+    '''
+    url = "https://mashape-community-urban-dictionary.p.rapidapi.com/define"
+    querystring = {"term": query}
+    headers = {
+        'x-rapidapi-key': os.getenv('urbanAPI'),
+        'x-rapidapi-host': "mashape-community-urban-dictionary.p.rapidapi.com"
+        }
+    response = requests.request("GET", url, headers=headers, params=querystring)
+    rjson = response.json()
+    e = discord.Embed(title = f"Urban Dictionary Meaning for: {query}", url = rjson['list'][0]['permalink'], color = random.randint(0, 16777215))
+    e.add_field(name = "Definition:", value = rjson['list'][0]['definition'], inline = False)
+    e.add_field(name = "Example:", value = rjson['list'][0]['example'], inline = False)
+    e.set_footer(text = f"👍: {rjson['list'][0]['thumbs_up']} | Author: {rjson['list'][0]['author']}")
+    await inter.send(embed = e)
+
 
   @commands.slash_command(name = "copy-person")
   @commands.bot_has_permissions(manage_webhooks = True)
